requirements.txt
# SmartVocab 智能词汇学习系统依赖包

# 数据库相关
mysql-connector-python==8.0.33

# 密码加密
bcrypt==4.0.1

# 深度学习相关
torch>=2.5.1
numpy>=1.21.0
pandas>=1.3.0

# Web框架 (用于API)
flask==2.3.2
# Flask 2.3 依赖 werkzeug 2.x；werkzeug 3+ 移除 __version__ 会导致 test_client 报错
werkzeug>=2.3.7,<3.0
flask-cors==4.0.0
flask-restful==0.3.10
gunicorn>=21.0.0

# 数据处理
scikit-learn>=1.0.0

# 工具库
python-dotenv==1.0.0
requests==2.31.0

# 开发工具
pytest==7.4.0
pytest-cov==4.1.0



